// src/pages/main-menu/order/user/menu/shop/component/orderForm/models/interfaces.ts
import {
  AccountInfo,
  CustomerInfo,
  SIMCardDetail,
} from "@/pages/main-menu/order/models/interfaces";

export interface ShopForm {
  acct?: AccountInfo;
  cust?: CustomerInfo;
  defLangId?: number;
  simCardId?: number;
  simCard?: SIMCardDetail;
}

// Generated by https://quicktype.io

export interface NumberDialogForm {
  accNbrStart?: string;
  rules?: string;
  terminationFlag?: string;
  accNbrEnd?: string;
  accNbrPriceStart?: number;
  accNbrPriceEnd?: number;
  nbrMaxQty?: string;
  accNbrClassId?: string;
  contactChannelId?: number;
  areaId?: string;
  orgId?: number;
  staffId?: number;
  indepProdSpecId?: number;
  custId?: number;
  servType?: number;
  spId?: number;
  bindFlag?: string;
  offerIds?: number[];
  chooseCount?: string;
  selectFlag?: string;
  reserveType?: string;
}

export interface AccNbrExDto {
  accNbrId?: number;
  prefix?: string;
  accNbr?: string;
  staffId?: number;
  orgId?: number;
  accNbrClassId?: number;
  accNbrPrice?: number;
  accNbrTypeId?: number;
  accNbrState?: string;
  hlrId?: number;
  areaId?: number;
  stateDate?: Date;
  comments?: string;
  preCharging?: number;
  accNbrClassName?: string;
  classComm?: string;
  pwd?: string;
  reserveDate?: Date;
  hlrName?: string;
  simCardId?: number;
  iccid?: string;
  simTypeName?: string;
  simTypeId?: string;
  reservePartyType?: string;
  reservePartyCode?: string;
  reserveType?: string;
  isBindingFlag?: string;
}

export interface ChooseNbrDto {
  accNbrId?: number;
  areaLimit?: number;
  orgLimit?: number;
  indepProdSpecId?: number;
  selectFlag?: string;
  opraType?: string;
  reserveType?: string;
  accNbrStart?: string;
  accNbrEnd?: string;
  accNbrClassId?: string;
  staffId?: number;
  custId?: number;
  prefix?: string;
  rules?: string;
  accNbrState?: string;
  stateFlag?: string;
  accNbr?: string;
  servType?: number;
  orgId?: number;
  areaId?: number;
  bindFlag?: string;
  iccidStart?: string;
  iccidEnd?: string;
  fromPPS?: boolean;
  accNbrTypeId?: number;
  hlrId?: number;
  simCardId?: number;
  chooseCount?: number;
  spId?: number;
  certTypeId?: number;
  certNbr?: string;
  pwd?: string;
  accNbrPriceStart?: number;
  accNbrPriceEnd?: number;
  terminationFlag?: string;
  sqlBuffer?: string; // StringBuffer -> string
  nbrMaxQty?: number;
}